Added monsters to the project but not fully to the scene.
Tried to update the movement mechanism to make it way smoother but it didn't work for now... it is back to the old one.

Removed "Equip" input action and it is intended to replace the key with a simple collision with the equitable item.

Added new scripts, "MobControl" and "MobSpawnLevel1" and they need more modification.

--ZeroNeroIV